Create README.md
Fixing a technical issue with Vysor on a Moto X v4 phone. I'm not an idiot (at least I don't think I am) but admit that technology support is so counterintuitive to me. So you guys are way ahead of me. I use Vysor to control my phone by voice through my PC because I'm paralyzed and can't use my hands. It worked very well for a long time. But now for months it has a "Connection Error" even after making sure it's plugged in and rebooting both the phone and computer. Most of the time after sending a text, it shows a connection error and mirroring disconnected. Then occasionally Vysor will flash a "Vysor is Connected" on my computer screen. But when I go back to the mirror app it has a black screen with an error message. Thanks for any help!